debezium/dbz#2101 Refactor HostPipelineController to use HostAllocation directly and extract HostDockerLogReader
